// projects/codx-ac/src/lib/models/AdvancedPaymentLines.model.ts
export class AdvancedPaymentLines{
    recID: any;
	transID: any;
    rowNo: any;
    reasonID: any;
    accountID: any;
    subLGType: any;
    objectType: any;
    objectID: any;
    settlement: any;
    settledID: any;
    settledNo: any;
    contractID: any;
    dr: any;
    cr: any;
    dR2: any;
    cR2: any;
    dim1: any;
    dim2: any;
    dim3: any;
    projectID: any;
    offsetAcctID: any;
    isBrigdeAcct: any;
    vatid: any;
    vatAmt: any;
    vatAmt2: any;
    illegalAmt: any;
    assetType: any;
    assetGroupID: any;
    serviceDate: any;
    servicePeriods: any;
    employeeID: any;
    siteID: any;
    refType: any;
    refNo: any;
    refLineID: any;
    note: any;
    singleEntry: any;
    autoCreated: any;
    createdOn: any;
    createdBy: any;
    modifiedOn: any;
    modifiedBy: any;
}